cmake: Only compile either static or shared depending on PPPLUGIN_SHARED
This change was done in order to remove the need for compiling the
object library with position independent code ("-fPIC").
Since a user will usually only need either the shared or the static
version of the library, only one of them can be compiled at a time.
If a user wants to compile and install both the static and shared
library, two separate builds are necessary.