include: Move include/ to include/ppplugin/
This is necessary to avoid users of the library having to add
include/ppplugin/ to the include path which may lead to confusing bugs
with common header names.
An alternative would be using a symlink include/ppplugin/ pointing to
include/, but a circular symlink may cause issues in some obscure build
system.