cmake: Fix boost find_package warning
CMake 3.30 removed the FindBoost module in favor of BoostConfig.cmake
which is included in boost v1.70 or newer. This new mechanism can be
used by simply adding "CONFIG" to the find_package call.
Alternatively, the minimum CMake version of 3.30 can be specified to
avoid the warning and explicitly use the backwards-compatible fallback
mechanism implemented in this CMake version.

Since many systems do not have CMake 3.30 yet, this commit upgrades
boost to v1.70 which is reasonably seasoned and should be available in
most cases and uses the new mechanism.