templates: Add SecondTemplateParameterOr helper
This will return the second template parameter of a template class.
It will be used for retrieving the error of Expected when it is not
clear that the class actually has enough template parameters.
A default value can be passed in case the class does not have sufficient
template parameters.