clang-tidy: Remove include-cleaner because of Python
Since Python does have its own philosophy for header includes (always
include Python.h as first header which also includes all other required
headers), it is not feasible to use the include-cleaner in this project.
Since there is no common pattern for the Python headers, it is also not
possible to simply use the option "IgnoreHeaders" to disable this check
only for Python headers.